// src/app/models/afiliado.ts
export class Afiliado {
    id: number;
    name: string;
    apellido: string;
    telefono: string;
    sexo: string;
    image_ine: string;
    estado_civil: string;
    fecha_nacimiento: Date;
    lugar_nacimiento: string;
    estado_vivienda: string;
    tiempo_viviendo: string;
    calle: string;
    colonia: string;
    dep_menores: boolean;
    dep_tercera_edad: boolean;
    Vivienda_compartida: boolean;
}